Moves "getPublicUrlFromRequest" to request utilities
